[TEST] handle dependancies for tests
