+++ setup.py
@@ -16,6 +16,14 @@ from setuptools import setup, find_packages
 requires = [
     'importlib_resources>=5.10.2,<5.11',
 ]
+extras_require = {
+    'tests': [
+        'pylint',
+        'pytest',
+        'pytest-cov',
+        'pytest-helpers_namespace'
+        ]
+}
 
 setup(
     name = 'narps_open',
@@ -27,9 +35,11 @@ setup(
     author = '',
     author_email = '',
     url = 'https://github.com/Inria-Empenn/narps_open_pipelines',
+
     include_package_data = True,
     python_requires = '>=3.8,<3.11',
     install_requires = requires,
+    extras_require = extras_require,
     classifiers = [
         'Programming Language :: Python',
         'Programming Language :: Python :: 3',
